refactor(ics): lift triggerICSDownload + slugifyGameName into lib
OverviewTabContent and ScheduleTabContent each defined the same
9-line downloadICS helper and rebuilt the same lowercase-hyphenate
slug. Both helpers now live in src/lib/ics.ts.
